Move gcov data handling to a own format handler
