+++ gcovr/formats/base.py
@@ -17,9 +17,9 @@
 #
 # ****************************************************************************
 
-from typing import List, Set
+from typing import List
 
-from ..options import GcovrConfigOption, Options, OutputOrDefault
+from ..options import GcovrConfigOption, Options
 from ..coverage import CovData
 
 
@@ -28,13 +28,15 @@ class handler_base(object):
         return []
 
     @staticmethod
-    def read_report(input_files: Set[str], options: Options) -> CovData:
-        raise RuntimeError(f"Function 'read_report' not implemented.")
+    def read_report(covdata: CovData, options: Options) -> bool:
+        raise RuntimeError("Function 'read_report' not implemented.")
 
     @staticmethod
     def write_report(covdata: CovData, output_file: str, options: Options) -> bool:
-        raise RuntimeError(f"Function 'write_report' not implemented.")
+        raise RuntimeError("Function 'write_report' not implemented.")
 
     @staticmethod
-    def write_summary_report(covdata: CovData, output_file: str, options: Options) -> bool:
-        raise RuntimeError(f"Function 'write_summary_report' not implemented.")
+    def write_summary_report(
+        covdata: CovData, output_file: str, options: Options
+    ) -> bool:
+        raise RuntimeError("Function 'write_summary_report' not implemented.")
